Rename database folder
